+++ packages/caelus/src/index.ts
@@ -16,3 +16,4 @@ export * from "./ephemeris.js";
 export * from "./features.js";
 export * from "./compiler.js";
 export * from "./lots.js";
+export * from "./profections.js";